+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog
 
 # Todo
-- rename repository from devenv to devenv2
 - add ci
 - none
 
 # v2023.1.1-beta
+- rename repository from devenv to devenv2
 - rename file lnsync.sh to devenv.sh